use random string instead of deleteAt for snapshot names

+++ fleetshard/pkg/central/cloudprovider/awsclient/rds_test.go
@@ -67,31 +67,32 @@ func waitForClusterToBeDeleted(ctx context.Context, rdsClient *RDS, clusterID st
 	}
 }
 
-func waitForFinalSnapshotToExist(ctx context.Context, rdsClient *RDS, snapshotID string) (bool, error) {
+func waitForFinalSnapshotToExist(ctx context.Context, rdsClient *RDS, clusterID string) (bool, string, error) {
 
 	ticker := time.NewTicker(awsRetrySeconds * time.Second)
 	for {
 		select {
 		case <-ticker.C:
 			snapshotOut, err := rdsClient.rdsClient.DescribeDBClusterSnapshots(&rds.DescribeDBClusterSnapshotsInput{
-				DBClusterSnapshotIdentifier: &snapshotID,
+				DBClusterIdentifier: &clusterID,
 			})
 
 			if err != nil {
 				if awsErr, ok := err.(awserr.Error); ok {
 					if awsErr.Code() != rds.ErrCodeDBClusterSnapshotNotFoundFault {
-						return false, err
+						return false, "", err
 					}
 
 					continue
 				}
 			}
 
-			if snapshotOut != nil {
-				return len(snapshotOut.DBClusterSnapshots) == 1, nil
+			if snapshotOut != nil && len(snapshotOut.DBClusterSnapshots) == 1 {
+				return true, *snapshotOut.DBClusterSnapshots[0].DBClusterSnapshotIdentifier, nil
 			}
+
 		case <-ctx.Done():
-			return false, fmt.Errorf("waiting for final DB snapshot: %w", ctx.Err())
+			return false, "", fmt.Errorf("waiting for final DB snapshot: %w", ctx.Err())
 		}
 
 	}
@@ -165,17 +166,18 @@ func TestRDSProvisioning(t *testing.T) {
 	require.NoError(t, err)
 	assert.True(t, clusterDeleted)
 
-	expectedSnapshotID := getFinalSnapshotID(clusterID, deletedAt)
-	// Always attempt to delete the final snapshot if it exists
-	defer func() {
-		_, err := rdsClient.rdsClient.DeleteDBClusterSnapshot(
-			&rds.DeleteDBClusterSnapshotInput{DBClusterSnapshotIdentifier: expectedSnapshotID},
-		)
+	snapshotExists, snapshotID, err := waitForFinalSnapshotToExist(deleteCtx, rdsClient, clusterID)
 
-		assert.NoError(t, err)
-	}()
+	if snapshotExists {
+		defer func() {
+			_, err := rdsClient.rdsClient.DeleteDBClusterSnapshot(
+				&rds.DeleteDBClusterSnapshotInput{DBClusterSnapshotIdentifier: &snapshotID},
+			)
 
-	snapshotExists, err := waitForFinalSnapshotToExist(deleteCtx, rdsClient, *expectedSnapshotID)
+			assert.NoError(t, err)
+		}()
+	}
+
 	require.NoError(t, err)
 	require.True(t, snapshotExists)
 }
